Added Makefila + migrated libs code to .cpp + checked lib build works
