+++ src/dynamic_linkage.hpp
@@ -1,0 +1,6 @@
+#ifndef DYNAMIC_HPP
+#define DYNAMIC_HPP
+
+void print_random_name();
+
+#endif // DYNAMIC_HPP
