+++ src/dynamic_load.hpp
@@ -1,0 +1,10 @@
+#ifndef DYNAMIC_LINKAGE_HPP
+#define DYNAMIC_LINKAGE_HPP
+
+#include <iostream>
+
+unsigned int power(unsigned int x, unsigned int p);
+
+void print_inversion(unsigned int x);
+
+#endif // DYNAMIC_LINKAGE_HPP
