+++ src/static.hpp
@@ -1,0 +1,6 @@
+#ifndef STATIC_HPP
+#define STATIC_HPP
+
+void print_ascii_number(unsigned int value);
+
+#endif // STATIC_HPP
